BC-1951 Add image browser command NOT WORKING
