+++ static/scripts/ckeditor/plugins/ckeditor5-filebrowser/src/filebrowser.js
@@ -6,37 +6,7 @@ import imageIcon from '@ckeditor/ckeditor5-core/theme/icons/image.svg';
 
 import audioIcon from '../theme/icons/audio.svg';
 import videoIcon from '../theme/icons/video.svg';
-
-const createFilebrowserModal = (editor, t, dialogTitle, onCreate, additionalInput = '') => {
-	const ckeditorFilebrowserDialog = $('.ckeditor-filebrowser-dialog');
-	populateModalForm(ckeditorFilebrowserDialog, {
-		title: dialogTitle,
-		closeLabel: t('Cancel'),
-		submitLabel: t('OK'),
-	});
-
-	const dialogContent = `<label for="url-input" style="display: none">${t('URL')}:</label>
-		<input type="hidden" id="url-input">
-		<input type="hidden" id="editor-id">
-		<button type="button" id="browseServerButton">${t('Browse Server')}</button><br>${additionalInput}`;
-
-	ckeditorFilebrowserDialog.find('.modal-body').html(dialogContent);
-	ckeditorFilebrowserDialog.find('.btn-submit').on('click', () => {
-		ckeditorFilebrowserDialog.modal('hide');
-		onCreate();
-	});
-	ckeditorFilebrowserDialog.appendTo('body').modal('show');
-
-	document.getElementById('editor-id').value = editor.id;
-	window.addEventListener('message', (e) => {
-		document.getElementById('url-input').value = e.data;
-	});
-
-	ckeditorFilebrowserDialog.find('#browseServerButton').on('click', () => {
-		const dialogPageUrl = `${editor.config.get('filebrowser.browseUrl')}?CKEditor=true`;
-		window.open(dialogPageUrl, '_blank', 'width=700, height=500');
-	});
-};
+import ImageBrowserCommand from './imagebrowsercommand';
 
 export default class FileBrowserPlugin extends Plugin {
 	init() {
@@ -70,33 +40,21 @@ export default class FileBrowserPlugin extends Plugin {
 		editor.ui.componentFactory.add('imagebrowser', (locale) => {
 			const view = new ButtonView(locale);
 
+			editor.commands.add('imagebrowser', new ImageBrowserCommand(editor));
+
+			const command = editor.commands.get('imagebrowser');
+			console.log('command is enabled: ', command.isEnabled);
+			console.log('command is enabled: ', command.value);
+
 			view.set({
 				label: t('Insert Image'),
 				icon: imageIcon,
 				tooltip: true,
+				command: 'imagebrowser',
 			});
 
-			view.on('execute', async () => {
-				const additionalInput = `<br><label for="alt-text-input">${t('Alternative Text')}:</label>
-				  <input type="text" id="alt-text-input">`;
-				const dialogTitle = t('Image Properties');
-				const onCreate = () => {
-					const imageUrl = document.getElementById('url-input').value;
-					if (!imageUrl) return;
-					const imageAltText = document.getElementById('alt-text-input').value;
-					editor.model.change((writer) => {
-						const imageElement = writer.createElement('image', {
-							src: imageUrl,
-							alt: imageAltText,
-						});
-						const lastOpenedEditorId = document.getElementById('editor-id').value;
-						if (lastOpenedEditorId === editor.id) {
-							editor.model.insertContent(imageElement, editor.model.document.selection);
-						}
-					});
-				};
-				createFilebrowserModal(editor, t, dialogTitle, onCreate, additionalInput);
-			});
+			view.bind('isEnabled').to(command);
+			console.log('command is enabled after: ', command.isEnabled);
 
 			return view;
 		});
